+++ llvm/docs/ReleaseNotes.md
@@ -180,6 +180,7 @@ Changes to LLDB
 * Support for FreeBSD on MIPS64 has been removed.
 * The minimum assumed version of FreeBSD is now 14. The effect of which is that watchpoints are
   assumed to be supported.
+* Support for FreeBSD on RISCV64 has been added.
 
 Changes to BOLT
 ---------------